requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-countries==7.3.2
django-debug-toolbar==3.6.0
django-discover-runner==1.0
django-filter==22.1
django-jazzmin==2.5.0
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
django-role-permissions==3.1.1
django-shortuuidfield==0.1.3
djangorestframework==3.13.1
haversine==2.7.0
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
mysqlclient==2.1.1
pycodestyle==2.9.1
python-decouple==3.6
pytz==2022.2.1
pyuca==1.2
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1
